% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{MQ_TMT_writeChannelAnnotationFile}
\alias{MQ_TMT_writeChannelAnnotationFile}
\title{Writes second annotation file to fill out channel, condition, bioReiplicate for each mixture.}
\usage{
MQ_TMT_writeChannelAnnotationFile(evidence_path, mix_annot_path, out_path)
}
\arguments{
\item{evidence_path}{input path to MaxQuant evidence file}

\item{mix_annot_path}{input path to mixture annotation file}

\item{out_path}{output path to channel annotation file}
}
\description{
Writes second annotation file to fill out channel, condition, bioReiplicate for each mixture.
}
\examples{
\dontrun{
MQ_TMT_writeChannelAnnotationFile("evidence.txt", "mixtureAnnotation.csv", "channelAnnotation.csv")
}

}
