% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{MQ_TMT_writeFullAnnotationFile}
\alias{MQ_TMT_writeFullAnnotationFile}
\title{Writes final annotation file that merges channel and run info}
\usage{
MQ_TMT_writeFullAnnotationFile(mix_annot_path, chan_annot_path, out_path)
}
\arguments{
\item{mix_annot_path}{input path to mixture annotation file}

\item{chan_annot_path}{input path to channel annotation file}

\item{out_path}{output path to final annotation file}
}
\description{
Writes final annotation file that merges channel and run info
}
\examples{
\dontrun{
MQ_TMT_writeFullAnnotationFile("mixtureAnnotation.csv", "channelAnnotation.csv", "annotation.csv")
}

}
