% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{MQ_TMT_writeMixtureAnnotationFile}
\alias{MQ_TMT_writeMixtureAnnotationFile}
\title{Writes first annotation file to fill out mixture and tech rep mixture.
Output is based on the header of the evidence file and the full summary file}
\usage{
MQ_TMT_writeMixtureAnnotationFile(evidence_path, summary_path, out_path)
}
\arguments{
\item{evidence_path}{input path to MaxQuant evidence file}

\item{summary_path}{input path to MaxQuant summary file}

\item{out_path}{output path to mixture annotation file}
}
\description{
Writes first annotation file to fill out mixture and tech rep mixture.
Output is based on the header of the evidence file and the full summary file
}
\examples{
\dontrun{
MQ_TMT_writeMixtureAnnotationFile("evidence.txt", "summary.txt", "mixtureAnnotation.csv")
}

}
